+++ src/test-fixtures/assert-xml/index.ts
@@ -1,0 +1,6 @@
+export {
+	assertExpectedElementQueries,
+	assertUnexpectedElementQueries,
+	createXmlAssertions,
+} from './assert-xml'
+export { splitXpathIntoSteps } from './xpath-parser'