+++ src/test-fixtures/index.ts
@@ -1,4 +1,4 @@
-
+export * from './assert-xml'
 export * from './config'
 export * from './constant'
 export * from './test-record'